% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dist.R
\name{fit_distribution}
\alias{fit_distribution}
\title{Fit distribution}
\usage{
fit_distribution(param)
}
\arguments{
\item{param}{rainfall parameter to fit (one of precipitation depth or time)}
}
\value{
two graphs illustrating the level of fit histogram with theoretical density function and empirical and theoretical CDFs, result table of fit tests and best parameters of fitted distribution.
}
\description{
Function \code{fit_distribution} fit one of Gamma, Log-normal, Pareto, Exponential, Weibull or GEV distribution to your vector.
}
\examples{
x <- rgamma(100, shape = 12, rate = 0.2)
fit_distribution(x)
x <- rlnorm(50, meanlog = 3, sdlog = 1)
fit_distribution(x)
}
