% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Fit logistic regression}
\usage{
logit(dt)
}
\arguments{
\item{dt}{data frame containing (0-1) dependent variable called \code{Overflow}}
}
\value{
function return logistic regression model
}
\description{
Function \code{logit} fit logistic model describing overflows based on precipitiation and catchment area parameters.
}
\examples{
data(dt.logit)
mod <- glm(Overflow~., data = dt.logit, family = binomial("logit"))
summary(mod)
}
