% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{mod.summary}
\alias{mod.summary}
\title{Summary of glm model}
\usage{
mod.summary(model)
}
\arguments{
\item{model}{glm model}
}
\value{
function return summary table of logistic regression
}
\description{
Summary table of glm model
}
\examples{
data(dt.logit)
mod <- glm(Overflow~., data = dt.logit, family = binomial("logit"))
mod.summary(mod)
}
