% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dist.R
\name{rain_dist}
\alias{rain_dist}
\title{Wrapper function which find the best fit distribution and parameters.}
\usage{
rain_dist(dt, param, output = "fit", rt = NULL)
}
\arguments{
\item{dt}{data frame}

\item{param}{precipitation parameter to fit}

\item{output}{\code{fit} (default) plots and table of fit test will be shown, or \code{best.fit} show the best fit distribution and parameters.}

\item{rt}{numeric vector of length 2, choose rainfall type setting lower and upper bound of rainfall event duration (your rainfall event duration variable in data frame should be named \code{t}).}
}
\value{
depending on \code{output} parameter plots and test table or best fit distribution will be shown
}
\description{
Function \code{rain_dist} allows you to choose which rain type you want to fit and output type.
}
\examples{
dt <- data.frame(P = rgamma(100, shape = 12, rate = 0.2), t = rlnorm(100, meanlog = 5, sdlog = 1))
rain_dist(dt, t, rt = c(0,150))
rain_dist(dt, t, output = "best.fit", rt = c(150, 630))
}
